pyproject.toml
[project]
name = "knowledge-graph-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "linkml>=1.9.6",
    "linkml-runtime>=1.9.5",
    "pyoxigraph>=0.5.4",
    "python-dotenv>=1.2.1",
    "rich>=14.3.1",
    "telethon>=1.42.0",
]
